lectures

+++ main.js
@@ -3,6 +3,7 @@ const client = new Discord.Client({ partials: ['MESSAGE', 'CHANNEL', 'REACTION']
 const CockroachDB = require('./database')
 
 const monitorTimeout = 30000
+const days = ['s', 'm', 't', 'w', 'th', 'f', 'sat']
 
 let db = new CockroachDB()
 
@@ -66,26 +67,58 @@ client.on('messageReactionRemove', async (reaction, user) => {
     }
 })
 
-function monitorAssignments(server) {
+function monitor(server) {
     if (server) {
         let notifChannel = server.channels.cache.find(chnl => chnl.name == 'alert-notifs')
 
         if (notifChannel) {
+            let date = new Date()
+            let currentDay = days[date.getDay()]
+            db.getLecture(server.id, (lecture) => {
+                if (lecture.days.includes(currentDay)) {
+                    let lectureTime = lecture.time.split(':')
+                    let timeLeft = ((parseInt(lectureTime[0]) * 60) + (parseInt(lectureTime[1]))) - ((date.getHours() * 60) + (date.getMinutes()) + (date.getSeconds() / 60))
+                    
+                    if (timeLeft <= 10.5 && timeLeft > 10) {
+                        const alertEmbed = new Discord.MessageEmbed()
+                        .setColor('FF3300')
+                        .setTitle('~Lecture Alert~')
+                        .setDescription('This lecture starts in 10 minutes, please join soon.')
+
+                        notifChannel.send(alertEmbed)
+                    }
+                        
+                }
+            })
+
             let now = Date.now()
             db.getAssignments(server.id, (assignments) => {
                 assignments.forEach(assignment => {
-                    console.log(timeLeft)
                     let timeLeft = assignment.due_date - now
-                    if (timeLeft <= (2.592e+8 + monitorTimeout) && timeLeft > 2.592e+8)
-                        notifChannel.send(`There are 3 days left before ${assignment.name} is due.`)
-                    else if (timeLeft <= (8.64e+7 + monitorTimeout) && timeLeft > 8.64e+7)
-                        notifChannel.send(`There is 1 day left before ${assignment.name} is due.`)
-                    else if (timeLeft <= (4.32e+7 + monitorTimeout) && timeLeft > 4.32e+7)
-                        notifChannel.send(`There are 12 hours left before ${assignment.name} is due!`)
-                    else if (timeLeft <= (2.16e+7 + monitorTimeout) && timeLeft > 2.16e+7)
-                        notifChannel.send(`There are 6 hours left before ${assignment.name} is due!!`)
-                    else if (timeLeft <= (3.6e+6 + monitorTimeout) && timeLeft > 3.6e+6)
-                        notifChannel.send(`There is 1 hour left before ${assignment.name} is due!!!`)
+                    const alertEmbed = new Discord.MessageEmbed()
+                        .setColor('FF3300')
+                        .setTitle('~Upcoming Deadline~')
+                    
+                    if (timeLeft <= (2.592e+8 + monitorTimeout) && timeLeft > 2.592e+8) {
+                        alertEmbed.addField(assignment.name, 'Due in 3 days.')
+                        notifChannel.send(alertEmbed)
+                    }
+                    else if (timeLeft <= (8.64e+7 + monitorTimeout) && timeLeft > 8.64e+7) {
+                        alertEmbed.addField(assignment.name, 'Due in 1 day.')
+                        notifChannel.send(alertEmbed)
+                    }
+                    else if (timeLeft <= (4.32e+7 + monitorTimeout) && timeLeft > 4.32e+7) {
+                        alertEmbed.addField(assignment.name, 'Due in 12 hours!')
+                        notifChannel.send(alertEmbed)
+                    }
+                    else if (timeLeft <= (2.16e+7 + monitorTimeout) && timeLeft > 2.16e+7) {
+                        alertEmbed.addField(assignment.name, 'Due in 6 hours!!')
+                        notifChannel.send(alertEmbed)
+                    }
+                    else if (timeLeft <= (3.6e+6 + monitorTimeout) && timeLeft > 3.6e+6) {
+                        alertEmbed.addField(assignment.name, 'Due in 1 hour!!!')
+                        notifChannel.send(alertEmbed)
+                    }
                 })
             })
         }
@@ -96,7 +129,8 @@ client.on('ready', () => {
     console.log(`Logged in as ${client.user.tag}!`)
     client.guilds.cache.forEach(async (server) => {
         try {
-            setInterval(monitorAssignments, monitorTimeout, server)
+            monitor(server)
+            setInterval(monitor, monitorTimeout, server)
 
             let welcomeChannel = server.channels.cache.find(chnl => chnl.name === 'welcome' && chnl.type === 'text')
 
@@ -110,7 +144,6 @@ client.on('ready', () => {
                     let roleEmbed = new Discord.MessageEmbed({
                         color: '#e87500',
                         description: 'Hello! Before you can view the server, you need to enter your name and select your pronouns. Please type your name and then react to your preferred pronouns.\n\n:one: - she/her\n:two: - he/him\n:three: - they/them\n:four: - they/she\n:five: - they/him\n:six: - any',
-                        
                     })
                     let msg = await welcomeChannel.send(roleEmbed)
                     await msg.react('1️⃣')
@@ -144,10 +177,7 @@ client.on('message', msg => {
         }
     }
 
-    //var command = msg.content.split(' ')
-    //return command[] = {!stuff, parameter}
-
-    if (welcomeChannels[msg.guild.id] === msg.channel.id) {
+    if (welcomeChannels[msg.guild.id] === msg.channel.id && !msg.author.bot) {
         msg.member.setNickname(msg.content)
         .then(() => {
             msg.delete() 
@@ -184,7 +214,7 @@ client.on('message', msg => {
             case '!add-assignment' :
                 let assignment = (msg.content.match(/\"(.*?)\"/))[1]
                 let count = 1 + assignment.split(' ').length
-                let dueDate = new Date(command[count], command[count + 1], command[count + 2])
+                let dueDate = new Date(command[count], command[count + 1]-1, command[count + 2])
                 dueDate.setTime(dueDate.getTime() + 86399000)
                 db.addAssignment(msg.guild.id, assignment, dueDate, (err, res) => {
                     if (err && err.code == 23505)
@@ -212,7 +242,7 @@ client.on('message', msg => {
                         let assignmentFields = []
                         assignments.forEach(a => {
                             let date = new Date(Number(a.due_date))
-                            assignmentFields.push({ name: a.name, value: `Due: ${date.getMonth()}/${date.getDate()}/${date.getFullYear()}`, inline: true })
+                            assignmentFields.push({ name: a.name, value: `Due: ${date.getMonth()+1}/${date.getDate()}/${date.getFullYear()}`, inline: true })
                         })
                         let assignmentEmbed = new Discord.MessageEmbed({
                             color: '#e87500',
@@ -236,11 +266,33 @@ client.on('message', msg => {
                     })
                     msg.member.roles.add(role)
                 }
+                break
+
+            case '!set-lecture':
+                let days = command[1].substr(1, command[1].length-2).split(',')
+                db.addLecture(msg.guild.id, days, convertTo24Hour(command[2]), (err, res) => {
+                    if (err)
+                        msg.reply('there was an issue setting the lecture, please try again.')
+                    else
+                        msg.reply('succesfully added the lecture date, students will now be notified before going live.')
+                })
+                break
 
             default:
                 break
         }
     }
 })
+
+function convertTo24Hour(time) {
+    var hours = parseInt((time.split(':'))[0]);
+    if(time.indexOf('am') != -1 && hours == 12) {
+        time = time.replace('12', '0');
+    }
+    if(time.indexOf('pm')  != -1 && hours < 12) {
+        time = time.replace(hours, (hours + 12));
+    }
+    return time.replace(/(am|pm)/, '') + ':00';
+}
 
 client.login('ODE1MjkwNDg3ODc1NDM2NTc2.YDqQbw.M8jUAUhHMHOT2rZ0M2aMt7fl_GY')
